Fixed background for AuthenticationScaffold. Wired with LoginScreen. Added onBackNavigation to LoginViewModel.
