+++ app/src/main/java/com/koleff/kare_android/ui/compose/components/navigation_components/scaffolds/AuthenticationScaffold.kt
@@ -1,8 +1,10 @@
 package com.koleff.kare_android.ui.compose.components.navigation_components.scaffolds
 
 import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -11,9 +13,15 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.drawWithContent
+import androidx.compose.ui.graphics.BlendMode
+import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
+import com.koleff.kare_android.R
 import com.koleff.kare_android.ui.compose.components.navigation_components.toolbar.AuthenticationToolbar
 import com.koleff.kare_android.ui.compose.shapes.RoundedToolbarShape
 
@@ -35,7 +43,7 @@ fun AuthenticationScaffold(
                     .fillMaxWidth()
                     .height(screenHeight / 2.5f)
                     .background(
-                        color = MaterialTheme.colorScheme.primaryContainer
+                        color = Color.Transparent
                     )
                     .border(
                         border = BorderStroke(2.dp, color = Color.White),
@@ -45,6 +53,34 @@ fun AuthenticationScaffold(
             )
         },
     ) { innerPadding ->
+
+        //Background image
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .drawWithContent {
+                    val colors = listOf(
+                        Color.Red,
+                        Color.Red,
+                        Color.Black,
+                        Color.Blue
+                    )
+                    drawContent()
+                    drawRect(
+                        brush = Brush.linearGradient(colors),
+                        blendMode = BlendMode.Overlay //ColorBurn
+                    )
+                }
+        ) {
+
+            //Texture background
+            Image(
+                painter = painterResource(id = R.drawable.ic_login_background_4),
+                contentDescription = "Background",
+                contentScale = ContentScale.Crop
+            )
+        }
+
         modifierPadding(innerPadding)
     }
 }
